+++ tests/test_create_tree.py
@@ -77,13 +77,11 @@ def test_add_variations():
             return MockJob()
 
     operations = []
-    test_variation = [
-        {
-            "operation": "n/a",
-            "schema": "n/a",
-            "values": [{"foo": 1}, {"foo": 2}, {"foo": 3}],
-        }
-    ]
+    test_variation = [{
+        "operation": "n/a",
+        "schema": "n/a",
+        "values": [{"foo": 1}, {"foo": 2}, {"foo": 3}],
+    }]
     id_path_mapping = {}
     operations = add_variations(
         operations, MockProject(), test_variation, MockJob(), id_path_mapping
@@ -93,7 +91,7 @@ def test_add_variations():
 
 
 def test_create_tree(tmpdir, emit_test_config):
-    project = OpenFOAMProject.init_project(root=tmpdir)
+    project = OpenFOAMProject.init_project(path=tmpdir)
 
     create_tree(project, emit_test_config, {"folder": tmpdir}, skip_foam_src_check=True)
 
@@ -120,7 +118,7 @@ def test_create_tree(tmpdir, emit_test_config):
 
 
 def test_call_generate_tree(tmpdir, emit_test_config):
-    project = OpenFOAMProject.init_project(root=tmpdir)
+    project = OpenFOAMProject.init_project(path=tmpdir)
     workspace_dir = tmpdir / "workspace"
     view_dir = tmpdir / "view"
 
@@ -148,7 +146,7 @@ def test_call_generate_tree(tmpdir, emit_test_config):
 
 def test_cache_folder(tmpdir, emit_test_config):
     emit_test_config["case"]["cache_folder"] = f"{tmpdir}/tmp"
-    project = OpenFOAMProject.init_project(root=tmpdir)
+    project = OpenFOAMProject.init_project(path=tmpdir)
 
     create_tree(project, emit_test_config, {"folder": tmpdir}, skip_foam_src_check=True)
 
@@ -168,8 +166,8 @@ def test_cache_folder(tmpdir, emit_test_config):
     # after purgin and recreating the workspace, the cache folder should be used
     Path(f"{tmpdir}/tmp/test").touch()
     shutil.rmtree(workspace_dir)
-    os.remove(f"{tmpdir}/signac.rc")
-    project = OpenFOAMProject.init_project(root=tmpdir)
+    shutil.rmtree(tmpdir /".signac")
+    project = OpenFOAMProject.init_project(path=tmpdir)
     create_tree(project, emit_test_config, {"folder": tmpdir}, skip_foam_src_check=True)
     project.run(names=["fetchCase"])