+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Changelog
 - Fix missing of groups in `obr run --list-operations` view, see https://github.com/hpsim/OBR/pull/159. 
 - Make view folders relative see https://github.com/hpsim/OBR/pull/164
 - Use cached version of git repo instead of cloning, see https://github.com/hpsim/OBR/pull/166
+- Bump signac version, see https://github.com/hpsim/OBR/pull/169
+- Validate simulation state after runSerial|ParallelSolver, see https://github.com/hpsim/OBR/pull/168
+
 
 0.2.0 (2023-09-14)
 ------------------